Separate metadata module
